+++ Source/NetOffice/COMDynamicObject.cs
@@ -1698,7 +1698,7 @@ namespace NetOffice
         /// <returns></returns>
         public static bool operator ==(COMDynamicObject objectA, COMObject objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1717,7 +1717,7 @@ namespace NetOffice
         /// <returns>true if equal, otherwise false</returns>
         public static bool operator !=(COMDynamicObject objectA, COMObject objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1736,7 +1736,7 @@ namespace NetOffice
         /// <returns></returns>
         public static bool operator ==(COMDynamicObject objectA, COMDynamicObject objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1755,7 +1755,7 @@ namespace NetOffice
         /// <returns></returns>
         public static bool operator ==(COMDynamicObject objectA, object objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1774,7 +1774,7 @@ namespace NetOffice
         /// <returns>true if equal, otherwise false</returns>
         public static bool operator ==(object objectA, COMDynamicObject objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1799,7 +1799,7 @@ namespace NetOffice
         /// <returns>true if equal, otherwise false</returns>
         public static bool operator !=(COMDynamicObject objectA, COMDynamicObject objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1818,7 +1818,7 @@ namespace NetOffice
         /// <returns>true if equal, otherwise false</returns>
         public static bool operator !=(COMDynamicObject objectA, object objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))
@@ -1837,7 +1837,7 @@ namespace NetOffice
         /// <returns>true if equal, otherwise false</returns>
         public static bool operator !=(object objectA, COMDynamicObject objectB)
         {
-            if (!Settings.Default.EnableOperatorOverlads)
+            if (!Settings.Default.EnableOperatorOverloads)
                 return Object.ReferenceEquals(objectA, objectB);
 
             if (Object.ReferenceEquals(objectA, null) && Object.ReferenceEquals(objectB, null))