+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.8.0-preview
+
+### Breaking Changes
+* `Settings.EnableOperatorOverlads` was renamed to `Settings.EnableOperatorOverloads` [#306](https://github.com/NetOfficeFw/NetOffice/issues/306)  
+  If you listened to changes using `Settings.PropertyChanged` event, ensure you fix your code with the new property name.
+
 ## v1.7.9
 * Links now point to Microsoft Docs website
 * Documented PowerPoint types related to animations and effects